add multi tissue correction to tortuosity constraint (#98)
* Add multi tissue correction to tortuosity constraint.

This commit adds the possibility to correct the tortuosity constraint
by taking into account the multi-tissue properties of the employed
model. To do this, T1_tortuosity is now a class and the new syntax of the
`set_tortuous_parameter` method of the `MultiCompartmentModel` class is
extended in such a way that it can take as input the S0 of the tissues
modelled by the intra-cellular and the extra-cellular compartments.
Backward compatibility is maintained.